teeny: take teeny's content for every path teeny owns
The six per-repo merges are three-way merges whose base is each repo's
rewritten main/teeny fork point, so git auto-merged main's 191 commits of
divergence into teeny's files wherever the two sides touched different
regions. That produced hybrids neither branch ever compiled -- spline.h, for
instance, ended up with main's runtime SplineVec work spliced into teeny's
copy, which teeny's pushpull/utils.h does not expect (it wants a `bounds`
member that main's spline::utils<Dynamic> does not have).

This branch is a consolidation of teeny, not a reconciliation of teeny with
main, so teeny's content is authoritative for every path teeny owns. Each of
the six rewritten teeny refs is checked back out over the merge result. The
build system is deliberately excluded: main's Makefile restructure is the
layout this branch adopts, and teeny's settings are carried into it separately.

Paths that exist only on main (solve_field, pushpull_backward, the CUDA
reg_*_rls split) are left as they are. They are not built here -- see the
MODULES and TESTEXCLUDE notes -- so the catch-up debt stays visible in one
place instead of being silently deleted.

+++ include/fastfields/api/cpu/restrict.h
@@ -1,6 +1,6 @@
 #ifndef FF_CPU_RESTRICT
 #define FF_CPU_RESTRICT
-#include "fastfields/core/dlpack.h"
+#include "dlpack.h"
 #include <cstdint>
 
 namespace ff  {
@@ -35,7 +35,7 @@ void restriction(
           double     shift  ,
     const double   * scale  ,
           int        ndim   ,
-          intptr_t   stream = 0
+          int        stream = 0
 );
 
 } // namespace cpu
