+++ include/fastfields/api/resize.h
@@ -1,8 +1,7 @@
 #ifndef FF_LIB_RESIZE
 #define FF_LIB_RESIZE
-#include "fastfields/core/dlpack.h"
-#include <cstdint>
-#include "fastfields/core/defines.h"
+#include "dlpack.h"
+#include "defines.h"
 
 #ifndef FF_LIB_BOUND_SPLINE_T
 #define FF_LIB_BOUND_SPLINE_T
@@ -64,7 +63,7 @@ void resample(
           double     shift  = 0.0,
     const double   * scale  = nullptr,
           int        ndim   = 1,
-          intptr_t   stream = 0
+          int        stream = 0
 );
 
 FF_NAMESPACE_END(FF)
